Fix mobile search modal interactions
