+++ web/src/lib/components/settings/ScheduledChatPickerDialog.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import * as Dialog from '$lib/components/ui/dialog';
-	import SidebarSearchDialog from '$lib/components/sidebar/SidebarSearchDialog.svelte';
+	import SidebarSearchPanel from '$lib/components/sidebar/SidebarSearchPanel.svelte';
 	import { matchesChatFilter, parseChatSearch } from '$shared/chat-filter-query';
 	import { getChatSessions } from '$lib/context';
 	import * as m from '$lib/paraglide/messages.js';
@@ -42,14 +42,13 @@
 	}
 </script>
 
-<Dialog.Root {open} onOpenChange={(value) => !value && onClose()}>
+<Dialog.Root {open} requestClose={onClose}>
 	<Dialog.Content
-		class="h-dvh w-screen max-w-none border-0 bg-transparent p-0 shadow-none sm:max-w-none"
+		class="top-[var(--app-viewport-center-y)] flex h-[var(--app-height)] max-h-[var(--app-height)] w-screen max-w-none flex-col gap-0 overflow-hidden rounded-none border-0 p-0 sm:w-screen sm:max-w-none min-[769px]:pointer-fine:top-[50%] min-[769px]:pointer-fine:h-[min(44rem,calc(var(--app-height)-2rem))] min-[769px]:pointer-fine:max-h-[44rem] min-[769px]:pointer-fine:w-[calc(100vw-2rem)] min-[769px]:pointer-fine:max-w-3xl min-[769px]:pointer-fine:rounded-2xl min-[769px]:pointer-fine:border"
 		showCloseButton={false}
 	>
 		<Dialog.Title class="sr-only">{m.scheduled_prompts_select_chat()}</Dialog.Title>
-		<SidebarSearchDialog
-			{open}
+		<SidebarSearchPanel
 			{query}
 			{filteredChats}
 			{currentTime}
@@ -66,9 +65,6 @@
 			onHighlightChange={(index) => (highlightedIndex = index)}
 			{onClose}
 			showSavedSearchActions={false}
-			overlayClass="z-[70]"
-			backdropTreatment="interaction-only"
-			contentRole="presentation"
 		/>
 	</Dialog.Content>
 </Dialog.Root>
