+++ web/src/lib/components/sidebar/Sidebar.svelte
@@ -1,12 +1,8 @@
 <script lang="ts">
-	import { onMount, untrack } from 'svelte';
+	import { untrack } from 'svelte';
 	import SidebarContent from './SidebarContent.svelte';
 	import SidebarSearchDock from './SidebarSearchDock.svelte';
 	import SidebarSelectionBar from './SidebarSelectionBar.svelte';
-	import SidebarSearchDialog from './SidebarSearchDialog.svelte';
-	import { searchResultNavigation } from '$lib/chat/actions/search-result-navigation.svelte.js';
-	import SavedSearchManagerDialog from './SavedSearchManagerDialog.svelte';
-	import SavedSearchEditorDialog from './SavedSearchEditorDialog.svelte';
 	import {
 		getAppShell,
 		getMinuteClock,
@@ -29,10 +25,6 @@
 	import { SidebarBulkDeleteState } from './sidebar-bulk-delete-state.svelte';
 	import { SidebarChatSelectionState } from '$lib/components/sidebar/sidebar-chat-selection-state.svelte.js';
 	import { addTagToQuery } from '$shared/chat-filter-query';
-	import {
-		EMPTY_TRANSCRIPT_SEARCH_INVALIDATION,
-		transcriptSearchInvalidationProjection,
-	} from '$lib/sidebar/search/transcript-search-invalidation.js';
 	import { buildSidebarDisplayChatIds, buildSidebarProjectKeys } from './sidebar-row-model';
 	import { SIDEBAR_SECTION_COLLAPSE_KEYS } from './sidebar-virtual-chat-list';
 	import {
@@ -44,13 +36,12 @@
 		SidebarChatItemLayout,
 		SidebarSortMode,
 	} from '$lib/stores/local-settings.svelte';
-	import type { ChatSearchSort } from '$shared/chat-search';
-	import type { SavedChatSearch } from '$lib/api/settings';
 	import * as Dialog from '$lib/components/ui/dialog';
 	import { Button } from '$lib/components/ui/button';
 	import * as m from '$lib/paraglide/messages.js';
 	import type { WorkspaceWindowEdge } from '$lib/workspace/surface-types.js';
 	import type { WorkspaceSplitAdmissions } from '$lib/workspace/window-geometry-policy.js';
+	import type { SavedChatSearch } from '$lib/api/settings';
 
 	interface QuickMoveWrite {
 		list: PersistedChatOrderGroup;
@@ -152,7 +143,6 @@
 	let isBulkOperating = $state(false);
 	let currentTime = $derived(minuteClock.currentTime);
 	let isMarkingAllRead = $state(false);
-	let transcriptSearchRetryVersion = $state(0);
 	let displayOptions = $derived<SidebarDisplayOptions>({
 		grouping: localSettings.sidebarGrouping,
 		inactivityDuration: localSettings.sidebarInactivityDuration,
@@ -161,25 +151,6 @@
 		sortMode: localSettings.sidebarSortMode,
 		pinnedInsertPosition: remoteSettings.snapshot?.ui?.pinnedInsertPosition ?? 'top',
 	});
-	let transcriptSearchTarget = $derived(
-		sidebarSearch.searchDialogOpen ? sidebarSearch.draftQuery : sidebarSearch.activeQuery,
-	);
-	let transcriptSearchEnabled = $derived(
-		remoteSettings.snapshot?.features?.transcriptSearch.enabled === true,
-	);
-	let transcriptSearchInvalidation = $derived.by(() => {
-		if (!transcriptSearchEnabled) return EMPTY_TRANSCRIPT_SEARCH_INVALIDATION;
-		return transcriptSearchInvalidationProjection(
-			chats,
-			transcriptSearchTarget,
-			localSettings.sidebarSearchResultSort,
-		);
-	});
-	let transcriptSearchHasTerms = $derived(transcriptSearchInvalidation.hasTranscriptTerms);
-	let transcriptSearchCandidateSet = $derived(transcriptSearchInvalidation.candidateSignature);
-	let transcriptSearchContentRevision = $derived(transcriptSearchInvalidation.contentSignature);
-	let transcriptSearchTimeOrder = $derived(transcriptSearchInvalidation.timeOrderSignature);
-
 	let visibleUnreadChatIds = $derived.by(() =>
 		sidebarSearch.filteredChats
 			.filter((chat) => chat.isUnread && Boolean(chat.lastActivityAt))
@@ -210,37 +181,6 @@
 		// preferences survive grouping-mode switches, like project keys do.
 		projectKeys.push(...SIDEBAR_SECTION_COLLAPSE_KEYS);
 		return projectKeys;
-	});
-
-	$effect(() => {
-		const query = transcriptSearchTarget;
-		const enabled = transcriptSearchEnabled;
-		const candidateSignature = transcriptSearchCandidateSet;
-		localSettings.sidebarSearchResultSort;
-		transcriptSearchRetryVersion;
-		untrack(() => sidebarSearch.updateTranscriptSearchCandidateSignature(candidateSignature));
-		if (!enabled || !transcriptSearchHasTerms) {
-			untrack(() => sidebarSearch.clearTranscriptSearch());
-			return;
-		}
-
-		const controller = new AbortController();
-		const timeoutId = setTimeout(() => {
-			void sidebarSearch.refreshTranscriptSearch(query, { signal: controller.signal });
-		}, 150);
-
-		return () => {
-			clearTimeout(timeoutId);
-			controller.abort();
-		};
-	});
-
-	$effect(() => {
-		const query = transcriptSearchTarget;
-		transcriptSearchContentRevision;
-		transcriptSearchTimeOrder;
-		if (query !== untrack(() => sidebarSearch.transcriptSearchQuery)) return;
-		untrack(() => sidebarSearch.scheduleTranscriptSearchRevalidation());
 	});
 
 	function handleChatClick(chatId: string) {
@@ -451,12 +391,6 @@
 		localSettings.set('sidebarSortMode', sortMode);
 	}
 
-	function handleSetSearchResultSort(sort: ChatSearchSort): void {
-		if (sort === localSettings.sidebarSearchResultSort) return;
-		localSettings.set('sidebarSearchResultSort', sort);
-		sidebarSearch.resetTranscriptSearchForSortChange();
-	}
-
 	function handleToggleChatListAutohide(): void {
 		if (!chatListAutohideAvailable) return;
 		const enabled = !localSettings.chatListAutohide;
@@ -466,20 +400,6 @@
 
 	function handleSetDockOnRight(enabled: boolean): void {
 		localSettings.set('chatListDock', enabled ? 'right' : 'left');
-	}
-
-	// Search dialog actions.
-
-	function handleSearchSelectChat(chatId: string) {
-		sidebarSearch.confirmSearchDialog();
-		void sidebarSearch.openTranscriptResult(chatId, (id, seq) => {
-			if (seq !== null) searchResultNavigation.set(id, seq);
-			onChatSelect(id);
-		});
-	}
-
-	function handleApplySavedSearch(search: SavedChatSearch) {
-		sidebarSearch.updateDraftQuery(search.query);
 	}
 
 	function handleApplySidebarMenuSearch(query: string) {
@@ -493,14 +413,6 @@
 	function handleClearActiveQuery() {
 		sidebarSearch.applyQuery('');
 	}
-
-	// Lifecycle.
-
-	onMount(() =>
-		appShell.onSidebarSearchRequested(() => {
-			sidebarSearch.toggleSearchDialog();
-		}),
-	);
 </script>
 
 <!-- The container delegates bubbled Escape handling for the sidebar subtree. Follow-up: CLEANUP_ROUND_TWO.md#a11y-suppression-register. -->
@@ -673,105 +585,6 @@
 				onclick={() => {
 					void confirmBulkDelete();
 				}}>{m.sidebar_actions_delete()}</Button
-			>
-		</Dialog.Footer>
-	</Dialog.Content>
-</Dialog.Root>
-
-<SidebarSearchDialog
-	open={sidebarSearch.searchDialogOpen}
-	portalToBody={isMobile}
-	query={sidebarSearch.draftQuery}
-	filteredChats={sidebarSearch.dialogDisplayChats}
-	savedSearches={sidebarSearch.searchDialogSavedSearches}
-	transcriptMatchesByChatId={sidebarSearch.transcriptSearchResultsByChatId}
-	{transcriptSearchEnabled}
-	transcriptSearchLoading={sidebarSearch.transcriptSearchLoading}
-	transcriptSearchIndexing={sidebarSearch.transcriptSearchIndexing}
-	transcriptSearchIndex={sidebarSearch.transcriptSearchIndex}
-	transcriptSearchStatus={sidebarSearch.transcriptSearchStatus}
-	transcriptSearchError={sidebarSearch.transcriptSearchError}
-	sort={localSettings.sidebarSearchResultSort}
-	showTranscriptPagination={sidebarSearch.transcriptSearchPage !== null}
-	hasMoreTranscriptResults={sidebarSearch.transcriptSearchPage?.hasMore === true &&
-		!sidebarSearch.transcriptSearchLimitReached}
-	loadingMoreTranscriptResults={sidebarSearch.transcriptSearchLoadingMore}
-	transcriptSearchPageError={sidebarSearch.transcriptSearchPageError}
-	transcriptSearchRevalidating={sidebarSearch.transcriptSearchRevalidating}
-	transcriptSearchRevalidationError={sidebarSearch.transcriptSearchRevalidationError}
-	transcriptSearchLimitReached={sidebarSearch.transcriptSearchLimitReached}
-	transcriptSearchAnnouncement={sidebarSearch.transcriptSearchAnnouncement}
-	transcriptSearchAnnouncementVersion={sidebarSearch.transcriptSearchAnnouncementVersion}
-	resultsResetVersion={sidebarSearch.transcriptSearchResultsResetVersion}
-	revalidationVersion={sidebarSearch.transcriptSearchRevalidationVersion}
-	{currentTime}
-	highlightedIndex={sidebarSearch.highlightedResultIndex}
-	onQueryChange={(q) => sidebarSearch.updateDraftQuery(q)}
-	onSelectChat={handleSearchSelectChat}
-	onApplySavedSearch={handleApplySavedSearch}
-	onOpenManager={() => sidebarSearch.openManagerFromSearchDialog()}
-	onCreateSavedSearch={() => sidebarSearch.openEditorForCreateFromSearchDialog()}
-	onHighlightChange={(i) => {
-		sidebarSearch.highlightedResultIndex = i;
-	}}
-	onRetryTranscriptSearch={() => {
-		transcriptSearchRetryVersion += 1;
-	}}
-	onSortChange={handleSetSearchResultSort}
-	onLoadMoreTranscriptResults={() => sidebarSearch.loadMoreTranscriptResults()}
-	onRetryTranscriptSearchRevalidation={() => sidebarSearch.retryTranscriptSearchRevalidation()}
-	reduceMotion={localSettings.reduceMotion}
-	onClose={() => sidebarSearch.closeSearchDialog()}
-/>
-
-<SavedSearchManagerDialog
-	open={sidebarSearch.managerOpen}
-	searches={sidebarSearch.savedSearches}
-	onClose={() => sidebarSearch.closeManager()}
-	onAdd={() => sidebarSearch.openEditorForCreate()}
-	onEdit={(search) => sidebarSearch.openEditorForEdit(search)}
-	onDelete={(id) => sidebarSearch.requestDelete(id)}
-	onReorder={(oldOrder, newOrder) => {
-		void sidebarSearch.reorder(oldOrder, newOrder);
-	}}
-/>
-
-<SavedSearchEditorDialog
-	editorState={sidebarSearch.editorState}
-	onClose={() => {
-		sidebarSearch.closeEditor();
-	}}
-	onSave={(data, searchId) => sidebarSearch.saveEditor(data, searchId)}
-/>
-
-<Dialog.Root
-	open={sidebarSearch.deleteConfirmation !== null}
-	onOpenChange={(open) => {
-		if (!open) sidebarSearch.clearDeleteConfirmation();
-	}}
->
-	<Dialog.Content
-		onOpenAutoFocus={(e) => {
-			e.preventDefault();
-			sidebarSearch.deleteButtonRef?.focus();
-		}}
-	>
-		<Dialog.Header>
-			<Dialog.Title>{m.sidebar_saved_searches_confirm_delete()}</Dialog.Title>
-			<Dialog.Description
-				>{m.sidebar_saved_searches_confirm_delete_description()}</Dialog.Description
-			>
-		</Dialog.Header>
-		<Dialog.Footer>
-			<Button variant="outline" onclick={() => sidebarSearch.clearDeleteConfirmation()}
-				>{m.sidebar_actions_cancel()}</Button
-			>
-			<Button
-				variant="destructive"
-				onclick={() => {
-					void sidebarSearch.confirmDelete();
-				}}
-				bind:ref={sidebarSearch.deleteButtonRef}>{m.sidebar_actions_delete()}</Button
 			>
 		</Dialog.Footer>
 	</Dialog.Content>
